Assignment:17.01
We are trying to analyze the results of a Portugese Bank marketing campaign conducted over phone. The goal of the campaign was to get a customer to sign up for a term deposit. The dataset collected is related to 17 campaigns that occurred between May 2008 and November 2010, corresponding to a total of 79354 contacts

We will follow the CRISP-DM process model consisting of Business Understanding, Data Understanding, Data Preparation, Modeling, Evaluation, Deployment stages.